backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
Flask-Login==0.6.3
email-validator==2.2.0
PyGithub==2.6.1
cryptography==42.0.8